Support retrieving the VM UUID on Windows
