+++ pkg/cloudprovider/providers/vsphere/vsphere_util_unsupported.go
@@ -1,0 +1,25 @@
+// +build !windows,!linux
+
+/*
+Copyright 2018 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package vsphere
+
+import "fmt"
+
+func GetVMUUID() (string, error) {
+	return "", fmt.Errorf("Retrieving VM UUID on this build is not implemented.")
+}
